coltype and table name options
table names:
you can now specify table names and what you want the new table to be
named

coltype: fixed it so that certain columns ( 60, 104, 106, 108) are cast
as varchar to get around the problem that some of these fields where
longer than what was in the database. I don’t know if this is the
smartest way around it but it worked.